Calendar summary partially working(average part)
The layout is pretty much complete. Missing the "No info"
message(eventually?) if no entry was found and the list of all the
reports(with the possibility of editing them by tapping).

Note that the CalendarView is not very friendly and should probably be
expanded with a button as "select" rather than using the listener on
date changes.

IMPORTANT: Fix rotation destroying everything.