Summary visualization stable
For a nice visualization of the report list a custom LinearLayout had to
be implemented(ListView had a problem since it's designed to have a
fixed size and make it scrollable for you).

The implemented LinearLayout(called AdaptableLinearLayout) has a
populateWithAdapter method which acts similarly to setAdapter for Views
that work with adapters.